Refactor in classes to fix compilations errors for new integrations
