+++ app/src/main/java/cy/agorise/bitsybitshareswallet/dao/CryptoNetAccountDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import cy.agorise.bitsybitshareswallet.enums.CryptoNetAccount
+import cy.agorise.bitsybitshareswallet.models.CryptoNetAccount
 
 @Dao
 interface CryptoNetAccountDao {
