+++ app/src/main/java/cy/agorise/bitsybitshareswallet/models/BitcoinAddress.kt
@@ -5,8 +5,6 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.ForeignKey.CASCADE
-import androidx.room.PrimaryKey
-import cy.agorise.bitsybitshareswallet.enums.CryptoNetAccount
 
 @Entity(tableName = "bitcoin_address",
     foreignKeys =  arrayOf(
